src/PythonService/requirements.txt
flask 
flask_sqlalchemy
pyodbc
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-sqlalchemy

# py -3 -m venv venv
# pip install -r requirements.txt


# Installare questa libreria
# https://visualstudio.microsoft.com/visual-cpp-build-tools/


# https://stackoverflow.com/questions/46405777/connect-docker-python-to-sql-server-with-pyodbc